Moving Switch data to its own file
This is in preparation for Switchsanity plando.
